third_party/WebKit/LayoutTests/fast/html/body-bgproperties.html

<!DOCTYPE html>
<title>The 'bgproperties=fixed' attribute on body</title>
<script src="../../resources/js-test.js"></script>
<body bgproperties=fixed>
<script>
shouldBeEqualToString("getComputedStyle(document.body).backgroundAttachment", "fixed");
</script>
