third_party/WebKit/LayoutTests/fast/html/body-bgproperties-expected.txt

CONSOLE WARNING: The 'bgproperties=fixed' attribute is deprecated. Please use CSS 'background-attachment:fixed' instead.
PASS getComputedStyle(document.body).backgroundAttachment is "fixed"
PASS successfullyParsed is true

TEST COMPLETE
